requirements.txt
dspy-ai>=2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
